+++ mod_redstoneExtended/src/net/minecraft/src/redstoneExtended/Laser/LaserShapes.java
@@ -8,8 +8,8 @@ public class LaserShapes {
     public static final LaserShape Bridge;
 
     static {
-        Default = new LaserShape(4f / 16f, false, (short)0, (byte)Block.blockSnow.blockIndexInTexture);
-        Deadly = new LaserShape(2f / 16f, false, (short)4, (byte)Block.blockSnow.blockIndexInTexture);
-        Bridge = new LaserShape(14f / 16f, true, (short)0, (byte)Block.glass.blockIndexInTexture);
+        Default = new LaserShape(4f / 16f, false, 0, Block.blockSnow.blockIndexInTexture);
+        Deadly = new LaserShape(2f / 16f, false, 4, Block.blockSnow.blockIndexInTexture);
+        Bridge = new LaserShape(14f / 16f, true, 0, Block.glass.blockIndexInTexture);
     }
 }
